- committed first set of API changes to allow reading / writing to L2 annotation - removed all layout specific code from SBML Document

+++ src/sbml/xml/XMLErrorLog.cpp
@@ -158,7 +158,7 @@ XMLErrorLog::getError (unsigned int n) const
 unsigned int
 XMLErrorLog::getNumErrors () const
 {
-  return mErrors.size();
+  return (unsigned int)mErrors.size();
 }
